Boot of Decks

Boot of Decks

Boot of Decks

Details

This R6 class simulates a boot of decks that allows drawing and reshuffling.

Public fields

deck

A list of DeckOfCards objects.

drawn

A data frame representing the drawn cards.

Methods

Public methods

BootOfDecks$new()

BootOfDecks$reshuffle()

BootOfDecks$draw()

BootOfDecks$solitaire()

BootOfDecks$clone()

Method new()

Initialize the boot of decks.

Usage

BootOfDecks$new(number_of_decks, use_jokers = FALSE)

Arguments

number_of_decks

The number of decks in the boot.

use_jokers

A logical value indicating whether to include jokers in the boot.

Returns

A list of DeckOfCards objects.

Method reshuffle()

Reshuffle the boot of decks.

Usage

BootOfDecks$reshuffle()

Returns

A data frame representing the reshuffled boot of decks.

Method draw()

Draw a specified number of cards from the boot.

Usage

BootOfDecks$draw(draws = 1)

Arguments

draws

The number of cards to draw from the boot.

Returns

A data frame representing the drawn cards.

Method solitaire()

sort the deck of cards.

Usage

BootOfDecks$solitaire(hand = "deck")

Arguments

hand

either the deck or the drawn hand.

Returns

a data frame representing the sorted deck of cards.

Method clone()

The objects of this class are cloneable with this method.

Usage

BootOfDecks$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Create a new boot of decks
my_boot <- BootOfDecks$new(2)
drawn_cards1 <- my_boot$draw(5)
# Print drawn cards
print(drawn_cards1)
#>     Card     Suit
#> 24  Jack Diamonds
#> 36    10    Clubs
#> 50  Jack   Spades
#> 291    3    Clubs
#> 171    4 Diamonds
# should be the same
drawn_cards1 == my_boot$drawn
#>     Card Suit
#> 24  TRUE TRUE
#> 36  TRUE TRUE
#> 50  TRUE TRUE
#> 291 TRUE TRUE
#> 171 TRUE TRUE
# Draw more cards
drawn_cards2 <- my_boot$draw(5)
# should be the same
rbind(drawn_cards1,drawn_cards2) == my_boot$drawn
#>     Card Suit
#> 24  TRUE TRUE
#> 36  TRUE TRUE
#> 50  TRUE TRUE
#> 291 TRUE TRUE
#> 171 TRUE TRUE
#> 30  TRUE TRUE
#> 511 TRUE TRUE
#> 19  TRUE TRUE
#> 451 TRUE TRUE
#> 351 TRUE TRUE
# Print remaining cards
print(my_boot$deck)
#>      Card     Suit
#> 23     10 Diamonds
#> 51  Queen   Spades
#> 42      3   Spades
#> 49     10   Spades
#> 8       8   Hearts
#> 5       5   Hearts
#> 41      2   Spades
#> 29      3    Clubs
#> 46      7   Spades
#> 21      8 Diamonds
#> 18      5 Diamonds
#> 45      6   Spades
#> 38  Queen    Clubs
#> 47      8   Spades
#> 44      5   Spades
#> 20      7 Diamonds
#> 6       6   Hearts
#> 15      2 Diamonds
#> 10     10   Hearts
#> 4       4   Hearts
#> 31      5    Clubs
#> 3       3   Hearts
#> 17      4 Diamonds
#> 13   King   Hearts
#> 27    Ace    Clubs
#> 39   King    Clubs
#> 37   Jack    Clubs
#> 35      9    Clubs
#> 14    Ace Diamonds
#> 22      9 Diamonds
#> 33      7    Clubs
#> 11   Jack   Hearts
#> 2       2   Hearts
#> 48      9   Spades
#> 32      6    Clubs
#> 52   King   Spades
#> 25  Queen Diamonds
#> 16      3 Diamonds
#> 1     Ace   Hearts
#> 12  Queen   Hearts
#> 7       7   Hearts
#> 34      8    Clubs
#> 40    Ace   Spades
#> 43      4   Spades
#> 28      2    Clubs
#> 9       9   Hearts
#> 26   King Diamonds
#> 231    10 Diamonds
#> 421     3   Spades
#> 491    10   Spades
#> 81      8   Hearts
#> 53      5   Hearts
#> 411     2   Spades
#> 461     7   Spades
#> 211     8 Diamonds
#> 181     5 Diamonds
#> 381 Queen    Clubs
#> 471     8   Spades
#> 441     5   Spades
#> 201     7 Diamonds
#> 61      6   Hearts
#> 151     2 Diamonds
#> 101    10   Hearts
#> 410     4   Hearts
#> 311     5    Clubs
#> 310     3   Hearts
#> 131  King   Hearts
#> 271   Ace    Clubs
#> 391  King    Clubs
#> 371  Jack    Clubs
#> 141   Ace Diamonds
#> 221     9 Diamonds
#> 331     7    Clubs
#> 111  Jack   Hearts
#> 210     2   Hearts
#> 481     9   Spades
#> 321     6    Clubs
#> 521  King   Spades
#> 251 Queen Diamonds
#> 161     3 Diamonds
#> 110   Ace   Hearts
#> 121 Queen   Hearts
#> 71      7   Hearts
#> 341     8    Clubs
#> 501  Jack   Spades
#> 191     6 Diamonds
#> 241  Jack Diamonds
#> 401   Ace   Spades
#> 301     4    Clubs
#> 431     4   Spades
#> 281     2    Clubs
#> 91      9   Hearts
#> 361    10    Clubs
#> 261  King Diamonds
# Reshuffle the boot
my_boot$reshuffle()
# should be empty
print(my_boot$drawn)
#> data frame with 0 columns and 0 rows