DeckOfCards

This R6 class simulates a deck of cards that allows drawing and reshuffling.

Details

Deck of Cards

This R6 class simulates a deck of cards that allows drawing and reshuffling.

Public fields

deck

A data frame representing the deck of cards.

drawn

A data frame representing the drawn cards.

Methods

Public methods

DeckOfCards$new()

DeckOfCards$draw()

DeckOfCards$reshuffle()

DeckOfCards$solitaire()

DeckOfCards$clone()

Method new()

Initialize the deck of cards.

Usage

DeckOfCards$new(use_jokers = FALSE)

Arguments

use_jokers

A logical value indicating whether to include jokers in the deck.

Returns

a data frame representing the deck of cards.

Method draw()

Draw a specified number of cards from the deck.

Usage

DeckOfCards$draw(draws = 1)

Arguments

draws

The number of cards to draw from the deck.

Returns

A data frame representing the drawn cards.

Method reshuffle()

Reshuffle the deck of cards.

Usage

DeckOfCards$reshuffle()

Returns

a data frame representing the reshuffled deck of cards.

Method solitaire()

sort the deck of cards.

Usage

DeckOfCards$solitaire(hand = "deck")

Arguments

hand

either the deck or the drawn hand.

Returns

a data frame representing the sorted deck of cards.

Method clone()

The objects of this class are cloneable with this method.

Usage

DeckOfCards$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Create a new deck
my_deck <- DeckOfCards$new()
# Shuffle the deck
my_deck$reshuffle()
# Draw cards
drawn_cards1 <- my_deck$draw(5)
# Print drawn cards
print(my_deck$drawn)
#>    Card   Suit
#> 4     4 Hearts
#> 1   Ace Hearts
#> 41    2 Spades
#> 50 Jack Spades
#> 29    3  Clubs
# should be the same
drawn_cards1 == my_deck$drawn
#>    Card Suit
#> 4  TRUE TRUE
#> 1  TRUE TRUE
#> 41 TRUE TRUE
#> 50 TRUE TRUE
#> 29 TRUE TRUE
# Draw more cards
drawn_cards2 <- my_deck$draw(3)
# should be the same
rbind(drawn_cards1, drawn_cards2) == my_deck$drawn
#>    Card Suit
#> 4  TRUE TRUE
#> 1  TRUE TRUE
#> 41 TRUE TRUE
#> 50 TRUE TRUE
#> 29 TRUE TRUE
#> 5  TRUE TRUE
#> 21 TRUE TRUE
#> 13 TRUE TRUE
# Print drawn cards
print(my_deck$drawn)
#>    Card     Suit
#> 4     4   Hearts
#> 1   Ace   Hearts
#> 41    2   Spades
#> 50 Jack   Spades
#> 29    3    Clubs
#> 5     5   Hearts
#> 21    8 Diamonds
#> 13 King   Hearts
# Check number of cards
nrow(rbind(drawn_cards1, drawn_cards2)) + nrow(my_deck$deck) == 52
#> [1] TRUE
# Reshuffle the deck
my_deck$reshuffle()
nrow(my_deck$deck) == 52
#> [1] TRUE
nrow(my_deck$drawn) == 0
#> [1] TRUE
# draw 40 cards
drawn_cards3 <- my_deck$draw(40)
# sort the remaining cards
my_deck$solitaire("deck") |> print()
#>     Card     Suit
#> 1    Ace   Hearts
#> 5      5   Hearts
#> 11  Jack   Hearts
#> 12 Queen   Hearts
#> 30     4    Clubs
#> 31     5    Clubs
#> 32     6    Clubs
#> 20     7 Diamonds
#> 23    10 Diamonds
#> 25 Queen Diamonds
#> 41     2   Spades
#> 42     3   Spades